Implemented Export Snaps in each snapcups.
