+++ app/src/components/adminConsole/currentCup/CurrentCupOptionsBar.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import Cup from "../../../types/Cup";
 import CurrentCupOpennessButton from "./CurrentCupOpennessButton";
 import CurrentCupPublishButton from "./CurrentCupPublishButton";
+import CurrentCupExportSnaps from "./CurrentCupExportSnaps";
 import { Entity } from "../../../types/Entity";
 
 const CurrentCupOptionsBar: React.FunctionComponent = (props: {
@@ -10,6 +11,7 @@ const CurrentCupOptionsBar: React.FunctionComponent = (props: {
 }) => {
     return (
         <>
+            <CurrentCupExportSnaps cup={props.cup} />
             <CurrentCupOpennessButton
                 cup={props.cup}
                 updateCups={props.updateCups}
